add log function

+++ sky/deploy/services.py
@@ -11,6 +11,7 @@ import shutil
 import datetime
 import chardet
 import traceback
+import logging
 from stat import ST_ATIME, ST_CTIME, ST_MTIME
 
 reload(sys)
@@ -23,30 +24,38 @@ DEPLOY_PATH = r"D:\backup\deploy"
 
 # 首先建立连接
 def connect(host):
-    host_ip = host.ip
-    host_user = host.host_user
-    password = host.password
+    try:
+        host_ip = host.ip
+        host_user = host.host_user
+        password = host.password
 
-    # 解密
-    decode_str = base64.decodestring(password)
-    password = decode_str[-1] + decode_str[-2] + decode_str[2:-2] + decode_str[1] + decode_str[0]
+        # 解密
+        decode_str = base64.decodestring(password)
+        password = decode_str[-1] + decode_str[-2] + decode_str[2:-2] + decode_str[1] + decode_str[0]
 
-    # 根据操作类型，升级文件的存放位置
-    os_id = host.os_type.os_id
-    if os_id == 1:
-        remote_path = r"\\" + host_ip + os.sep + "d$"
-    else:
-        remote_path = r"\\" + host_ip + os.sep + host_user
+        # 根据操作类型，升级文件的存放位置
+        os_id = host.os_type.os_id
+        if os_id == 1:
+            remote_path = r"\\" + host_ip + os.sep + "d$"
+        else:
+            remote_path = r"\\" + host_ip + os.sep + host_user
 
-    cmd = r"net use %s %s /user:%s" % (remote_path, password, host_user)
+        cmd = r"net use %s %s /user:%s" % (remote_path, password, host_user)
 
-    flag = os.system(cmd)
-    log_str = ""
-    if flag == 0:
-        print u"连接成功"
-        log_str = datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + u" 连接成功" + "\n"
-    else:
-        print u"连接失败，请检查。。。"
+        flag = os.system(cmd)
+        log_str = ""
+        if flag == 0:
+            print u"连接成功"
+            log_str = datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + u" 连接成功" + "\n"
+        else:
+            print u"连接失败，请检查。。。"
+            log_str = datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + u" 连接失败，请检查。。。" + "\n"
+
+    except Exception as e:
+        traceback.print_exc()
+        logging.error("connection has an error:")
+        logging.exception(Exception)
+        flag = -1
         log_str = datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + u" 连接失败，请检查。。。" + "\n"
 
     return flag, log_str
@@ -67,8 +76,9 @@ def copy_deploy_file(order, host):
         if not os.path.exists(remote_path):
             os.makedirs(remote_path)
         deploy_file = os.path.join(DEPLOY_PATH, order_code+".zip")
-
-        shutil.copy(deploy_file, remote_path)
+        remote_path_file = os.path.join(remote_path, order_code+".zip")
+        if not os.path.exists(remote_path_file):
+            shutil.copy(deploy_file, remote_path)
 
         print u"复制成功"
         flag = 0
@@ -76,8 +86,9 @@ def copy_deploy_file(order, host):
 
     except Exception as e:
         traceback.print_exc()
-        print u"复制失败，请检查。。。"
-        flag = 1
+        logging.error(order_code+" copy has an error:")
+        logging.exception(Exception)
+        flag = -1
         log_str = datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + u" 复制失败，请检查。。。" + "\n"
 
     return flag, log_str
@@ -105,8 +116,9 @@ def deploy_backup(order, host):
         log_str = datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + u" 备份成功" + "\n"
     except Exception as e:
         traceback.print_exc()
-        print u"备份失败，请检查。。。"
-        flag = 1
+        logging.error(order_code+" backup has an error:")
+        logging.exception(Exception)
+        flag = -1
         log_str = datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + u" 备份失败，请检查。。。" + "\n"
 
     return flag, log_str
@@ -149,14 +161,22 @@ def deploy_do(order, host):
         # 删除文件
         shutil.rmtree(deploy_file_path)
 
-        print u"发布成功"
-        flag = 0
-        log_str = datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + u" 发布成功" + "\n"
+        if "复制了".encode("gb2312") in "".join(result):
+            print u"发布成功"
+            logging.info(order_code + u" has success flag 复制了")
+            flag = 0
+            log_str = datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + u" 发布成功" + "\n"
+        else:
+            print u"发布失败"
+            logging.info(order_code + u" has no success flag 复制了")
+            flag = -1
+            log_str = datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + u" 发布失败，请检查。。。" + "\n"
 
     except Exception as e:
         traceback.print_exc()
-        print u"发布失败，请检查。。。"
-        flag = 1
+        logging.error(order_code+" deploy has an error:")
+        logging.exception("EXCEPTION")
+        flag = -1
         log_str = datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + u" 发布失败，请检查。。。" + "\n"
 
     return flag, log_str
@@ -174,68 +194,74 @@ def generate_file_md5value(file_path):
 
 # md5验证，服务器端与远端发布后文件进行校验
 def md5_check(order, host):
-    order_code = order.order_code
-    module = order.module
-    program_path = module.program_path
-    host_ip = host.ip
+    try:
+        order_code = order.order_code
+        module = order.module
+        program_path = module.program_path
+        host_ip = host.ip
 
-    remote_path = r"\\" + host_ip + "\\" + program_path
-    deploy_file_zip = os.path.join(DEPLOY_PATH, order_code + ".zip")
+        remote_path = r"\\" + host_ip + "\\" + program_path
+        deploy_file_zip = os.path.join(DEPLOY_PATH, order_code + ".zip")
 
-    file_unzip = zipfile.ZipFile(deploy_file_zip, 'r')
-    for file_ in file_unzip.namelist():
-        file_unzip.extract(file_, DEPLOY_PATH)
-    # 获取解压后根目录
-    app_root_dir = file_unzip.namelist()[0][:-1]
-    deploy_file_path = os.path.join(DEPLOY_PATH, app_root_dir)
+        file_unzip = zipfile.ZipFile(deploy_file_zip, 'r')
+        for file_ in file_unzip.namelist():
+            file_unzip.extract(file_, DEPLOY_PATH)
+        # 获取解压后根目录
+        app_root_dir = file_unzip.namelist()[0][:-1]
+        deploy_file_path = os.path.join(DEPLOY_PATH, app_root_dir)
 
-    # 先建立连接
-    flag, log_str = connect(host)
+        # 先建立连接
+        flag, log_str = connect(host)
 
-    md5_form_list = []
-    if flag == 0:
-        remote_file_dict = {}
-        for root, dirs, files in os.walk(remote_path):
-            for file_ in files:
-                file_path = os.path.join(root, file_)
-                # 保证文件路径也是一样的，防止同名
-                key = file_path.replace(remote_path, "").upper()
-                remote_file_dict[key] = file_path
+        md5_form_list = []
+        if flag == 0:
+            remote_file_dict = {}
+            for root, dirs, files in os.walk(remote_path):
+                for file_ in files:
+                    file_path = os.path.join(root, file_)
+                    # 保证文件路径也是一样的，防止同名
+                    key = file_path.replace(remote_path, "").upper()
+                    remote_file_dict[key] = file_path
 
-        # 汇总检查结果，1：表示检查无不同
-        result_all = 1
-        for root, dirs, files in os.walk(deploy_file_path):
-            for file_ in files:
-                deploy_file = os.path.join(root, file_)
-                # 保证文件路径也是一样的，防止同名
-                key = deploy_file.replace(deploy_file_path, "").upper()
-                md5_form = dict()
+            # 汇总检查结果，1：表示检查无不同
+            result_all = 1
+            for root, dirs, files in os.walk(deploy_file_path):
+                for file_ in files:
+                    deploy_file = os.path.join(root, file_)
+                    # 保证文件路径也是一样的，防止同名
+                    key = deploy_file.replace(deploy_file_path, "").upper()
+                    md5_form = dict()
 
-                md5_form["deploy_file"] = deploy_file.replace(DEPLOY_PATH + os.sep, "")
-                md5_form["md5_source"] = generate_file_md5value(deploy_file)
-                if key in remote_file_dict.keys():
-                    remote_file = remote_file_dict[key]
-                    md5_form["remote_file"] = remote_file
-                    md5_form["md5_remote"] = generate_file_md5value(remote_file)
-                    if md5_form["md5_source"] == md5_form["md5_remote"]:
-                        md5_form["check_result"] = 1
+                    md5_form["deploy_file"] = deploy_file.replace(DEPLOY_PATH + os.sep, "")
+                    md5_form["md5_source"] = generate_file_md5value(deploy_file)
+                    if key in remote_file_dict.keys():
+                        remote_file = remote_file_dict[key]
+                        md5_form["remote_file"] = remote_file
+                        md5_form["md5_remote"] = generate_file_md5value(remote_file)
+                        if md5_form["md5_source"] == md5_form["md5_remote"]:
+                            md5_form["check_result"] = 1
+                        else:
+                            md5_form["check_result"] = 0
+                            # 只要有一个不同就标识异常
+                            result_all = 0
                     else:
+                        md5_form["remote_file"] = "can not find the file"
+                        md5_form["md5_remote"] = "has no md5"
                         md5_form["check_result"] = 0
                         # 只要有一个不同就标识异常
                         result_all = 0
-                else:
-                    md5_form["remote_file"] = "can not find the file"
-                    md5_form["md5_remote"] = "has no md5"
-                    md5_form["check_result"] = 0
-                    # 只要有一个不同就标识异常
-                    result_all = 0
-                md5_form_list.append(md5_form)
+                    md5_form_list.append(md5_form)
 
-    else:
-        print u"连接失败，请检查。。。"
+        else:
+            print u"连接失败，请检查。。。"
 
-    # 删除文件
-    shutil.rmtree(deploy_file_path)
+        # 删除文件
+        shutil.rmtree(deploy_file_path)
+    except Exception as e:
+        traceback.print_exc()
+        logging.error("md5 check has an error:")
+        logging.exception("EXCEPTION")
+
     return md5_form_list, result_all
 
 
@@ -264,32 +290,39 @@ def deploy_zip(order, host):
 
 # 发布Sql
 def deploy_sql(order, host):
-    order_code = order.order_code
-    host_ip = host.ip
-    host_user = host.host_user
-    password = host.password
+    try:
+        order_code = order.order_code
+        host_ip = host.ip
+        host_user = host.host_user
+        password = host.password
 
-    # 解密
-    decode_str = base64.decodestring(password)
-    password = decode_str[-1] + decode_str[-2] + decode_str[2:-2] + decode_str[1] + decode_str[0]
+        # 解密
+        decode_str = base64.decodestring(password)
+        password = decode_str[-1] + decode_str[-2] + decode_str[2:-2] + decode_str[1] + decode_str[0]
 
-    log_str = ""
-    flag = 0
-    order_deploy_file_path = os.path.join(DEPLOY_PATH, str(order_code))
-    for root, divs, files in os.walk(order_deploy_file_path):
-        for file_ in files:
-            if file_.endswith(".sql") or file_.endswith(".SQL"):
-                file_sql = os.path.join(root, file_.decode("gb2312"))
-                cmd = r"sqlcmd -S %s -U %s -P %s -i %s"%(host_ip, host_user, password, file_sql)
-                result = ".".join(os.popen(cmd.encode("gb2312")).readlines())
-                if "错误".encode("gb2312") in result or "级别".encode("gb2312") in result:
-                    flag = 1
-                    log_str = log_str + datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + " " + file_ + "\n" + result
-                    print log_str
-                    break
-                else:
-                    log_str = log_str + datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + " " + file_ + "\n" + result
-                    print log_str
+        log_str = ""
+        flag = 0
+        order_deploy_file_path = os.path.join(DEPLOY_PATH, str(order_code))
+        for root, divs, files in os.walk(order_deploy_file_path):
+            for file_ in files:
+                if file_.endswith(".sql") or file_.endswith(".SQL"):
+                    file_sql = os.path.join(root, file_)
+                    cmd = r"sqlcmd -S %s -U %s -P %s -i %s"%(host_ip, host_user, password, file_sql.decode("gb2312"))
+                    result = ".".join(os.popen(cmd.encode("gb2312")).readlines())
+                    if "错误".encode("gb2312") in result or "级别".encode("gb2312") in result:
+                        flag = 1
+                        log_str = log_str + datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + " " + file_ + "\n" + result
+                        print log_str
+                        break
+                    else:
+                        log_str = log_str + datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + " " + file_ + "\n" + result
+                        print log_str
+    except Exception as e:
+        traceback.print_exc()
+        logging.error(order_code+" has an error:")
+        logging.exception(Exception)
+        flag = 1
+        log_str = log_str + datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + " " + file_ + "\n" + result
 
     return flag, log_str.decode("gb2312")
 
@@ -358,8 +391,9 @@ def rollback(order, host):
 
     except Exception as e:
         traceback.print_exc()
-        print u"回滚失败，请检查。。。"
-        flag = 1
+        logging.error(order_code+"rollback has an error:")
+        logging.exception(Exception)
+        flag = -1
         log_all = log_all + datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S") + u" 回滚失败，请检查。。。" + "\n"
 
     return flag, log_all
